Add the Ja Rule Bootloader ID.
Per the DFU 1.1 spec:

... to ensure that only the DFU driver is loaded, it is considered necessary
to change the idProduct field of the device when it enumerates the DFU
descriptor set.

This document does not attempt to specify how a vendor might alter the device’s
product ID except to suggest that adding one, setting the high bit, or using
FFFFh are all valid possibilities. Vendors may use any scheme that they choose.

I went with adding 1 to the Ja Rule product ID.