SoloKeys PID code application
For our open source FIDO2 security token, we have an official
vid/pid of 0483/a2ca from STMicroelectronics. We use this for
the consumer product, which is locked down for safety reasons,
allowing only firmware updates signed by us via our bootloader.

We would like to set a separate PID of 50B0 for this bootloader.

For enthusiasts, we offer the "hacker" edition of the key, which
allows arbitrary (unsigned or self-signed) firmware to be loaded.
It would be great to use a separate PID of 5070 for this.

Please let me know if any modifications are needed (in particular,
I used GPL-3.0-or-later, and didn't see a way to set the hardware
license in the individual PID applications).

Thanks a lot for this service,
Nicolas