Clarify reserved PIDs in the howto
Previously, this page would refer to the VID page for the reserved PIDs,
but also list the 1xxx reservation explicitly while omitting the 0xxx
reservation. This commit lists both, but more briefly, removing some
duplication.

This also changes the reference "VID page" to "PID list page", since
that matches the link at the start of the paragraph (from the link text
"list of PIDs" it is not immediately obvious that it links to the VID
page).