Request for a PID for PewPew LCD device.
PewPew is a family of handheld devices running CircuitPython designed
for use in programming workshops. PewPew LCD is a new device in this
family, using a monochrome LCD display.

The VID/PID is required to be able to merge the board definition for
this device into the CircuitPython repository. The pull request for
that is at https://github.com/adafruit/circuitpython/pull/6331